Move containers to components
